requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
pytz>=2023.3

# LLM Integration (OpenRouter via OpenAI SDK)
openai>=1.12.0

# Data & Analysis
vnstock>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ta>=0.11.0

# CLI
rich>=13.7.0
click>=8.1.0

# Testing
pytest>=7.4.0
